3.3

+++ src/main/java/hh/sof3/bookstore/domain/BookRepository.java
@@ -2,6 +2,6 @@ package hh.sof3.bookstore.domain;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface BookRepository extends CrudRepository<Book, Integer> {
+public interface BookRepository extends CrudRepository<Book, Long> {
 
 }
